text_summarize_api/requirements.txt
azure-functions
azure-storage-blob
pandas
pyarrow
fastparquet

